bugfix: make it compile with gcc 4.6

+++ Framework/Platforms/Source/OpenCVImageLoader/OpenCVImageLoader.h
@@ -94,7 +94,7 @@ private:
   //vector containing all file names in directory
   vector<string> allFiles;
   //current position
-  unsigned int currentPos;
+  size_t currentPos;
 
   GameData theGameData;
   unsigned int time;
